ATAGG/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-WTF==1.2.1
python-slugify==8.0.1
Pillow==10.1.0
Flask-Session==0.5.0